sources: Fix build on LLP64 systems
We want to avoid, for example, where pointer values are getting
truncated as pointers are 8 bytes in size but sizeof(long) is only 4
bytes.

+++ src/cairo-script-surface.c
@@ -3065,7 +3065,7 @@ _emit_scaled_glyph_vector (cairo_script_surface_t *surface,
     cairo_script_context_t *ctx = to_context (surface);
     cairo_script_implicit_context_t old_cr;
     cairo_status_t status;
-    unsigned long index;
+    size_t index;
 
     index = ++font_private->subset_glyph_index;
     scaled_glyph->dev_private_key = ctx;
@@ -3113,7 +3113,7 @@ _emit_scaled_glyph_bitmap (cairo_script_surface_t *surface,
 {
     cairo_script_context_t *ctx = to_context (surface);
     cairo_status_t status;
-    unsigned long index;
+    size_t index;
 
     index = ++font_private->subset_glyph_index;
     scaled_glyph->dev_private_key = ctx;
